+++ LootGoblin/Models/BotState.cs
@@ -1,0 +1,18 @@
+namespace LootGoblin.Models;
+
+public enum BotState
+{
+    Idle,
+    SelectingMap,
+    OpeningMap,
+    DetectingLocation,
+    Teleporting,
+    Mounting,
+    WaitingForParty,
+    Flying,
+    OpeningChest,
+    InCombat,
+    InDungeon,
+    Completed,
+    Error,
+}